lab 17 - added conditional statement

+++ Jenkinsfile
@@ -1,5 +1,5 @@
 pipeline {
-    agent { label "agentfarm" }
+    agent { label 'agentfarm' }
     stages {
         stage('Delete the workspace') {
       steps {
@@ -8,8 +8,16 @@ pipeline {
         }
         stage('Installing Ansible') {
       steps {
-        sh 'sudo apt-get update -y && sudo apt-get upgrade -y'
-        sh 'sudo apt install -y wget tree unzip ansible python3-pip python3-apt'
+        script {
+            def ansible_exists = fileExists'/usr/bin/ansible'
+
+            if (ansible_exists) {
+                echo 'Ansible is already installed'
+            } else {
+                sh 'sudo apt-get update -y && sudo apt-get upgrade -y'
+                sh 'sudo apt install -y wget tree unzip ansible python3-pip python3-apt'
+            }
+        }
       }
         }
         stage('Third Stage') {
